Solution architecture sorted
