+++ ShopApp/Areas/Admin/Views/_ViewStart.cshtml
@@ -1,0 +1,3 @@
+﻿@{
+    Layout = "~/Views/Shared/_Layout.cshtml";
+}
