+++ ShopApp/Areas/Customer/Views/Home/Privacy.cshtml
@@ -1,0 +1,6 @@
+﻿@{
+    ViewData["Title"] = "Privacy Policy";
+}
+<h1>@ViewData["Title"]</h1>
+
+<p>Use this page to detail your site's privacy policy.</p>
