+++ ShopApp/Areas/Customer/Views/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+﻿@using ShopApp
+@using ShopApp.Models
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
